Move phantom-related code to another file
