+++ lib/compile-to-html.js
@@ -1,9 +1,8 @@
 var Hapi = require('hapi')
 var Inert = require('inert')
 var Path = require('path')
-var Phantom = require('phantomjs-prebuilt')
-var ChildProcess = require('child_process')
 var PortFinder = require('portfinder')
+var PhantomBrowser = require('browsers/phantom')
 
 module.exports = function (staticDir, route, options, callback) {
   function serveAndPrerenderRoute () {
@@ -53,42 +52,16 @@ module.exports = function (staticDir, route, options, callback) {
           // If port is already bound, try again with another port
           if (error) return serveAndPrerenderRoute()
 
-          var maxAttempts = options.maxAttempts || 5
-          var attemptsSoFar = 0
-
-          var phantomArguments = [
-            Path.join(__dirname, 'phantom-page-render.js'),
-            'http://localhost:' + port + route,
-            JSON.stringify(options)
-          ]
-
-          if (options.phantomOptions) {
-            phantomArguments.unshift(options.phantomOptions)
+          function callback (error, output) {
+            if (error) throw error
+            callback(output)
+            Server.stop()
           }
 
-          function capturePage () {
-            attemptsSoFar += 1
-
-            ChildProcess.execFile(
-              Phantom.path,
-              phantomArguments,
-              {maxBuffer: 1048576},
-              function (error, stdout, stderr) {
-                if (error || stderr) {
-                  // Retry if we haven't reached the max number of capture attempts
-                  if (attemptsSoFar <= maxAttempts) {
-                    return capturePage()
-                  } else {
-                    if (error) throw error
-                    if (stderr) throw stderr
-                  }
-                }
-                callback(stdout)
-                Server.stop()
-              }
-            )
+          if (options.browser === 'chrome') {
+          } else {
+            PhantomBrowser(options, port, route, callback)
           }
-          capturePage()
         })
       })
     })
